+++ src/snowflake/cli/api/commands/flags.py
@@ -134,11 +134,12 @@ AuthenticatorOption = typer.Option(
 
 PrivateKeyPathOption = typer.Option(
     None,
+    "--private-key-file",
     "--private-key-path",
-    help="Snowflake private key path. Overrides the value specified for the connection.",
+    help="Snowflake private key file path. Overrides the value specified for the connection.",
     hide_input=True,
     callback=_callback(
-        lambda: get_cli_context_manager().connection_context.set_private_key_path
+        lambda: get_cli_context_manager().connection_context.set_private_key_file
     ),
     show_default=False,
     rich_help_panel=_CONNECTION_SECTION,
@@ -267,18 +268,17 @@ _DIAG_LOG_DEFAULT_VALUE = "<temporary_directory>"
 
 
 def _diag_log_path_callback(path: str):
-    if path != _DIAG_LOG_DEFAULT_VALUE:
-        return path
-    return tempfile.gettempdir()
+    if path == _DIAG_LOG_DEFAULT_VALUE:
+        path = tempfile.gettempdir()
+    get_cli_context_manager().connection_context.set_diag_log_path(Path(path))
+    return path
 
 
 DiagLogPathOption: Path = typer.Option(
     _DIAG_LOG_DEFAULT_VALUE,
     "--diag-log-path",
     help="Diagnostic report path",
-    callback=_callback(
-        lambda: get_cli_context_manager().connection_context.set_diag_log_path
-    ),
+    callback=_diag_log_path_callback,
     show_default=False,
     rich_help_panel=_CONNECTION_SECTION,
     exists=True,