+++ src/snowflake/cli/api/entities/common.py
@@ -8,6 +8,8 @@ from snowflake.cli.api.sql_execution import SqlExecutor
 class EntityActions(str, Enum):
     BUNDLE = "action_bundle"
     DEPLOY = "action_deploy"
+    DROP = "action_drop"
+    VALIDATE = "action_validate"
 
 
 T = TypeVar("T")