+++ src/snowflake/cli/api/entities/utils.py
@@ -32,6 +32,7 @@ from snowflake.cli.api.errno import (
     DOES_NOT_EXIST_OR_CANNOT_BE_PERFORMED,
     NO_WAREHOUSE_SELECTED_IN_SESSION,
 )
+from snowflake.cli.api.exceptions import SnowflakeSQLExecutionError
 from snowflake.cli.api.project.schemas.entities.common import PostDeployHook
 from snowflake.cli.api.project.util import unquote_identifier
 from snowflake.cli.api.rendering.sql_templates import (
@@ -302,27 +303,34 @@ def render_script_templates(
     - List of rendered scripts content
     Size of the return list is the same as the size of the input scripts list.
     """
-    scripts_contents = []
-    for relpath in scripts:
-        script_full_path = SecurePath(project_root) / relpath
-        try:
-            template_content = script_full_path.read_text(file_size_limit_mb=UNLIMITED)
-            env = override_env or choose_sql_jinja_env_based_on_template_syntax(
-                template_content, reference_name=relpath
-            )
-            result = env.from_string(template_content).render(jinja_context)
-            scripts_contents.append(result)
+    return [
+        render_script_template(project_root, jinja_context, script, override_env)
+        for script in scripts
+    ]
 
-        except FileNotFoundError as e:
-            raise MissingScriptError(relpath) from e
 
-        except jinja2.TemplateSyntaxError as e:
-            raise InvalidTemplateInFileError(relpath, e, e.lineno) from e
+def render_script_template(
+    project_root: Path,
+    jinja_context: dict[str, Any],
+    script: str,
+    override_env: Optional[jinja2.Environment] = None,
+) -> str:
+    script_full_path = SecurePath(project_root) / script
+    try:
+        template_content = script_full_path.read_text(file_size_limit_mb=UNLIMITED)
+        env = override_env or choose_sql_jinja_env_based_on_template_syntax(
+            template_content, reference_name=script
+        )
+        return env.from_string(template_content).render(jinja_context)
 
-        except jinja2.UndefinedError as e:
-            raise InvalidTemplateInFileError(relpath, e) from e
+    except FileNotFoundError as e:
+        raise MissingScriptError(script) from e
 
-    return scripts_contents
+    except jinja2.TemplateSyntaxError as e:
+        raise InvalidTemplateInFileError(script, e, e.lineno) from e
+
+    except jinja2.UndefinedError as e:
+        raise InvalidTemplateInFileError(script, e) from e
 
 
 def validation_item_to_str(item: dict[str, str | int]):
@@ -330,3 +338,27 @@ def validation_item_to_str(item: dict[str, str | int]):
     if item["errorCode"]:
         s = f"{s} (error code {item['errorCode']})"
     return s
+
+
+def drop_generic_object(
+    console: AbstractConsole,
+    object_type: str,
+    object_name: str,
+    role: str,
+    cascade: bool = False,
+):
+    """
+    Drop object using the given role.
+    """
+    sql_executor = get_sql_executor()
+    with sql_executor.use_role(role):
+        console.step(f"Dropping {object_type} {object_name} now.")
+        drop_query = f"drop {object_type} {object_name}"
+        if cascade:
+            drop_query += " cascade"
+        try:
+            sql_executor.execute_query(drop_query)
+        except:
+            raise SnowflakeSQLExecutionError(drop_query)
+
+        console.message(f"Dropped {object_type} {object_name} successfully.")