+++ tests_integration/test_snowpark.py
@@ -870,7 +870,7 @@ def test_ignore_anaconda_uses_version_from_zip(
 
 @pytest.mark.integration
 def test_incorrect_requirements(project_directory, runner, alter_requirements_txt):
-    from pkg_resources._vendor.packaging.requirements import InvalidRequirement
+    from packaging.requirements import InvalidRequirement
 
     with project_directory("snowpark") as tmp_dir:
         alter_requirements_txt(